+++ src/components/textarea/textarea.stories.tsx
@@ -4,6 +4,7 @@ import { ComponentStory } from "@storybook/react";
 import CarbonProvider from "../carbon-provider/carbon-provider.component";
 
 import Textarea from ".";
+import I18nProvider from "../i18n-provider";
 
 export const DefaultStory: ComponentStory<typeof Textarea> = () => {
   const [state, setState] = useState("");
@@ -63,7 +64,7 @@ export const CharacterLimitStory: ComponentStory<typeof Textarea> = () => {
       expandable
       value={value}
       onChange={({ target }) => setValue(target.value)}
-      characterLimit="50"
+      characterLimit={50}
     />
   );
 };
@@ -78,14 +79,46 @@ export const UnenforcedCharacterLimitStory: ComponentStory<
       expandable
       value={value}
       onChange={({ target }) => setValue(target.value)}
-      characterLimit="50"
+      characterLimit={50}
       enforceCharacterLimit={false}
-      warnOverLimit
     />
   );
 };
 UnenforcedCharacterLimitStory.parameters = {
   chromatic: { disableSnapshot: true },
+};
+
+export const TranslationsCharacterLimitStory: ComponentStory<
+  typeof Textarea
+> = () => {
+  const [value, setValue] = useState("");
+  return (
+    <I18nProvider
+      locale={{
+        locale: () => "fr-FR",
+        characterCount: {
+          hintString: () => "L'entrée contient un compteur de caractères",
+          tooManyCharacters: (count, formattedCount) =>
+            count === 1
+              ? `Vous avez ${formattedCount} personnage de trop`
+              : `Vous avez ${formattedCount} personnages de trop`,
+          charactersLeft: (count, formattedCount) =>
+            count === 1
+              ? `Il vous reste ${formattedCount} personnage`
+              : `Il vous reste ${formattedCount} personnages`,
+        },
+      }}
+    >
+      <Textarea
+        label="Textarea"
+        expandable
+        value={value}
+        onChange={({ target }) => setValue(target.value)}
+        characterLimit={50}
+        enforceCharacterLimit={false}
+      />
+    </I18nProvider>
+  );
 };
 
 export const LabelInlineStory: ComponentStory<typeof Textarea> = () => {